15-Text2SQL自助式数据报表开发/SQL-vanna/requirements.txt
mysql_connector_repackaged==0.3.1
openai==1.77.0
vanna==0.7.9
pysqlite3==0.5.4
chromadb
PyMySQL
